+++ src/components/CurrentImage/index.css
@@ -9,9 +9,6 @@
   display: flex;
   flex-direction: column;
   background: #f0f0f0;
-  cursor: url('https://uploads.codesandbox.io/uploads/user/b3e56831-8b98-4fee-b941-0e27f39883ab/Ad1_-cursor.png') 39 39,
-    auto;
-
 }
 
 .selectedTile {
