--- src/components/GameSpace/index.css
@@ -1,11 +1,0 @@
-.puzzlePage__gameSpace {
-    background-color: #b4d0c4;
-    height: 100vh; /* use of vh is temporary until the game content is ready */
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.puzzlePage__puzzleBox {
-    background-color: #3b8c79;
-}
