+++ src/components/ListOfImages/index.jsx
@@ -1,9 +1,17 @@
 import React from "react";
+import {Row, Col} from "antd"
 
 const ListOfImages = () => {
 
+    // The cols in the return state are going to contain a list of clickable images 
+    // each image when clicked will show the puzzle of that specific image
     return (
-        <div></div>
+        <Row>
+            <Col>something</Col>
+            <Col>something</Col>
+            <Col>something</Col>
+            <Col>something</Col>
+        </Row>
     )
 }
